pyproject.toml
[tool.coverage.run]
omit = ["backend/tests/*"]

[tool.coverage.report]
fail_under = 80
skip_empty = true

[tool.coverage.html]
directory = ".tests_coverage"

[tool.ruff]
preview = true
line-length = 120
fix = true
output-format = "grouped"

[tool.ruff.format]
exclude = [".git", ".venv", ".mypy_cache", ".tox", "__pycache__"]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
warn_unused_ignores = true
strict_equality = true
extra_checks = true
warn_unused_configs = true
warn_redundant_casts = true

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "testinsight-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "cryptography>=45.0.6",
  "faker>=37.5.3",
  "fastapi>=0.116.1",
  "freezegun>=1.5.5",
  "fuzzysearch>=0.8.0",
  "gitpython>=3.1.43",
  "google-genai>=1.30.0",
  "httpx>=0.28.1",
  "lxml>=6.0.0",
  "pre-commit>=4.3.0",
  "pydantic>=2.11.7",
  "python-dotenv>=1.1.1",
  "python-jenkins>=1.8.3",
  "python-multipart>=0.0.20",
  "requests>=2.32.4",
  "tox>=4.27.0",
  "uvicorn>=0.35.0",
]

[dependency-groups]
dev = ["ipdb>=0.13.13", "ipython>=9.4.0"]
tests = [
  "pytest>=8.4.1",
  "pytest-asyncio>=1.1.0",
  "pytest-cov>=6.2.1",
  "pytest-mock>=3.14.1",
]

[tool.setuptools.packages.find]
include = ["backend*"]
exclude = ["tests*", "frontend*", "data*"]
